batch update

+++ FormsApp/Controllers/HomeController.cs
@@ -147,13 +147,39 @@ namespace FormsApp.Controllers
             {
                 return NotFound();
             }
-            Repository.DeleteProduct(entity);
-            return RedirectToAction("Index");
+            return View("DeleteConfirm", entity);
 
 
 
         }
+        [HttpPost]
+        public IActionResult Delete(int id, int ProductId)
+
+        {
+            if (id != ProductId)
+            {
+                return NotFound();
+            }
+            var entity = Repository.Products.FirstOrDefault(p => p.ProductId == id);
+            if (entity == null)
+            {
+                return NotFound();
+            }
 
 
+            Repository.DeleteProduct(entity);
+            return RedirectToAction("Index");
+
+
+        }
+        [HttpPost]
+        public IActionResult EditProducts (List<Product> Products)
+        {
+            foreach (var product in Products)
+            {
+                Repository.EditIsActive(product);
+            }
+            return RedirectToAction("Index");
+        }
     }
 }
